+++ R/Information.R
@@ -44,8 +44,12 @@ Log2TreesMatchingSplit <- function (A, B = A[2]) {
 
 #' Character information content
 #'
-#' `CharacterInformation()` calculates the phylogenetic information content
-#' of a given character.
+#' `CharacterInformation()` calculates the cladistic information content
+#' \insertCite{Steel2006}{TreeTools} of a given character, in bits.
+#' The total information in all characters gives a measure of the potential
+#' utility of a dataset \insertCite{Cotton2008}{TreeTools}, which can be
+#' compared with a profile parsimony score \insertCite{Faith2001}{TreeTools} to
+#' evaluate the degree of homoplasy within a dataset.
 #'
 #' @param tokens Character vector specifying the tokens assigned to each taxon for
 #' a character.  Example: `c(0, 0, 0, 1, 1, 1, '?', '-')`.
@@ -58,7 +62,7 @@ Log2TreesMatchingSplit <- function (A, B = A[2]) {
 #' 2006), in bits.
 #'
 #' @references
-#' - \insertRef{Steel2006}{TreeTools}
+#' - \insertAllCited{}
 #' @family split information functions
 #' @template MRS
 #' @importFrom fastmatch %fin%
@@ -75,9 +79,9 @@ CharacterInformation <- function (tokens) {
 
 #' Phylogenetic information content of splitting leaves into two partitions
 #'
-#' Calculate the phylogenetic information content (_sensu_
-#' Steel & Penny, 2006) of a split, which reflects the probability that a
-#' uniformly selected random tree will contain the split:
+#' Calculate the phylogenetic information content
+#' \insertCite{@sensu @Steel2006}{TreeTools} of a split, which reflects the
+#' probability that a uniformly selected random tree will contain the split:
 #' a split that is consistent with a smaller number of trees will have a higher
 #' information content.
 #'
@@ -116,7 +120,7 @@ CharacterInformation <- function (tokens) {
 #'
 #' Summing the information content of all splits within a tree, perhaps using
 #' the '[TreeDist](https://ms609.github.io/TreeDist/)' function
-#' [`SplitwiseInfo()`](https://ms609.github.io/TreeDist/reference/SplitwiseInfo.html),
+#' [`SplitwiseInfo()`](https://ms609.github.io/TreeDist/reference/TreeInfo.html),
 #' arguably gives a more instructive picture of its resolution than simply
 #' counting the number of splits that are present -- though with the caveat
 #' that splits within a tree are not independent of one another, so some
@@ -128,7 +132,7 @@ CharacterInformation <- function (tokens) {
 #' [`QuartetStates()`](https://ms609.github.io/Quartet/reference/QuartetState.html),
 #' or to use a different take on the information contained within a split, the
 #' clustering information: see the 'TreeDist' function
-#' [`ClusteringInfo()`](https://ms609.github.io/TreeDist/reference/ClusteringEntropy.html)
+#' [`ClusteringInfo()`](https://ms609.github.io/TreeDist/reference/TreeInfo.html)
 #' for details.
 #'
 #'
@@ -149,19 +153,18 @@ CharacterInformation <- function (tokens) {
 #' # that seven maximally uneven splits on the same leaves:
 #' SplitInformation(25, 25)
 #' 7 * SplitInformation(2, 48)
-#' @references
-#' - \insertRef{Steel2006}{TreeTools}
+#' @references \insertAllCited{}
 #'
 #' @family split information functions
 #'
 #' @seealso
 #'
 #' Sum the phylogenetic information content of splits within a tree:
-#'  [`TreeDist::SplitwiseInfo()`](https://ms609.github.io/TreeDist/reference/SplitwiseInfo.html)
+#'  [`TreeDist::SplitwiseInfo()`](https://ms609.github.io/TreeDist/reference/TreeInfo.html)
 #'
 #'
 #' Sum the clustering information content of splits within a tree:
-#'  [`TreeDist::ClusteringInfo()`](https://ms609.github.io/TreeDist/reference/ClusteringEntropy.html)
+#'  [`TreeDist::ClusteringInfo()`](https://ms609.github.io/TreeDist/reference/TreeInfo.html)
 #'
 #'
 #'
@@ -189,7 +192,8 @@ MultiSplitInformation <- function (partitionSizes) {
 #' Number of trees consistent with split
 #'
 #' Calculates the number of unrooted bifurcating trees consistent with the
-#' specified multi-partition split, using the formula of Carter _et al_. (1990).
+#' specified multi-partition split, using theorem two of
+#' \insertCite{Carter1990;textual}{TreeTools}.
 #'
 #' @param \dots A series or vector of integers listing the number of tips in
 #' each of a number of tree splits (e.g. bipartitions).
@@ -201,13 +205,11 @@ MultiSplitInformation <- function (partitionSizes) {
 #'  number of unrooted bifurcating trees consistent with the specified split.
 #'
 #'
+#' @references \insertAllCited{}
+#' 
 #' @examples
-#'  UnrootedTreesMatchingSplit(c(3, 5))
-#'  UnrootedTreesMatchingSplit(3, 2, 1, 2)
-#'
-#' @references
-#' See Theorem 2 in \insertRef{Carter1990}{TreeTools}
-#'
+#' UnrootedTreesMatchingSplit(c(3, 5))
+#' UnrootedTreesMatchingSplit(3, 2, 1, 2)
 #' @template MRS
 #' @family split information functions
 #' @export