+++ R/Splits.R
@@ -40,7 +40,6 @@
 #' moreSplits %in% splits
 #'
 #' @family Splits operations
-#' @importFrom ape reorder.phylo
 #' @name Splits
 #' @export
 as.Splits <- function (x, tipLabels = NULL, ...) UseMethod('as.Splits')
@@ -378,13 +377,19 @@ c.Splits <- function (...) {
 length.Splits <- function (x) nrow(x)
 
 #' @family Splits operations
+#' @importFrom stats setNames
 #' @export
 duplicated.Splits <- function (x, incomparables = FALSE, ...) {
   dupX <- !x
   useOrig <- x[, 1] < dupX[, 1]
   dupX[useOrig, ] <- x[useOrig, ]
-
-  duplicated.array(dupX, MARGIN = 1, ...)
+  
+  if (dim(dupX)[2] == 1) {
+    setNames(duplicated.default(dupX, incomparables = incomparables, ...),
+             names(dupX))
+  } else {
+    duplicated.array(dupX, MARGIN = 1, incomparables = incomparables, ...)
+  }
 }
 
 #' @family Splits operations