+++ man-roxygen/datasetParam.R
@@ -1,2 +1,2 @@
-#' @param dataset A phylogenetic data matrix of class \code{\link[phangorn]{phyDat}},
+#' @param dataset A phylogenetic data matrix of class `phyDat`,
 #'  whose names correspond to the labels of any accompanying tree.