+++ man-roxygen/edgeLengthParam.R
@@ -1,0 +1,3 @@
+#' @param edgeLength Numeric specifying edge lengths of consensus tree;
+#' `NULL` aligns tips by scaling edges proportional to clade size; `1` sets
+#' all edges to unit length.