+++ tests/testthat/test-tree_rearrange.R
@@ -6,15 +6,23 @@ nasty <- structure(list(edge = structure(
   tip.label = letters[1:8]),
   class = 'phylo') # Danger: Do not plot!
 
+test_that(".DescendantTips() recurses", {
+  nTip <- 19L
+  tree <- BalancedTree(nTip)
+  edge <- tree$edge
+  expect_equal(.DescendantTips(edge[, 1], edge[, 2], nTip, c(23, 32, 35)),
+               c(1:3, 11:13, 16:19))
+})
 
 test_that("RootOnNode() works", {
+  expect_null(RootOnNode(NULL, 1))
 
   tree <- structure(list(edge = structure(c(6L, 9L, 9L, 7L, 7L, 8L, 8L,
                                             6L, 9L, 2L, 7L, 3L, 8L, 4L, 5L, 1L),
                                           .Dim = c(8L, 2L)),
                          tip.label = c("t3", "t4", "t1", "t2", "t5"),
                          Nnode = 4L), class = "phylo", order = "cladewise")
-
+  
   exp8 <- structure(list(edge = structure(c(6L, 7L, 8L, 8L, 7L, 6L, 9L, 9L, 7L,
                                             8L, 1L, 2L, 3L, 9L, 4L, 5L),
                                           .Dim = c(8L, 2L)),
@@ -109,11 +117,10 @@ test_that("RootOnNode() works", {
   expect_true(all.equal(UnrootTree(PectinateTree(8)),
                         RootOnNode(PectinateTree(8), 9L, FALSE)))
   expect_true(all.equal(urt, RootOnNode(urt, 9L, FALSE)))
-  expect_true(all.equal(Preorder(EnforceOutgroup(urt, letters[1:2])),
+  expect_true(all.equal(Preorder(RootTree(urt, letters[1:2])),
                         RootOnNode(urt, 9L, TRUE)))
 
 })
-
 
 test_that("root_on_node() works", {
   tree <- Preorder(BalancedTree(15))
@@ -170,14 +177,24 @@ test_that("RootOnNode() supports nasty node ordering", {
                RootOnNode(nasty, 13L))
 })
 
+test_that("RootTree() handles null outgroups", {
+  bal8 <- BalancedTree(8)
+  expect_equal(bal8, RootTree(bal8))
+  expect_equal(bal8$edge, RootTree(bal8$edge))
+  expect_equal(bal8, RootTree(bal8, NULL))
+  expect_equal(bal8$edge, RootTree(bal8$edge, NULL))
+  expect_equal(bal8, RootTree(bal8, character(0)))
+  expect_null(RootTree(NULL))
+})
+
 test_that("RootTree() works", {
   bal8 <- BalancedTree(8)
+  bal15 <- BalancedTree(15)
   expect_error(RootTree(bal8, 1:8 %in% 0))
-  expect_error(RootTree(bal8, character(0)))
-  expect_error(RootTree(bal8, integer(0)))
   expect_error(RootTree(bal8, 'tip_not_there'))
   expect_equal(RootTree(bal8, 5:6), RootTree(bal8, 1:8 %in% 5:6))
   expect_equal(RootTree(bal8$edge, 5:6), RootTree(bal8, 5:6)$edge)
+  expect_equal(RootTree(bal15$edge, 9:11), RootTree(bal15, 9:11)$edge)
   expect_equal(RootTree(bal8, 5:6), RootTree(bal8, c('t5', 't6')))
   expect_true(all.equal(
                as.phylo(5518, 8, paste0('t', rev(c(7,8,3,4,1,2,6,5)))),
@@ -213,6 +230,7 @@ test_that("RootTree() works", {
 })
 
 test_that("UnrootTree() works", {
+  expect_null(UnrootTree(NULL))
   expect_equal(matrix(c(7, 8, 8, 7, 7, 9, 10, 10, 9,
                         8, 1, 2, 3, 9, 10, 4,  5, 6), ncol = 2L),
                UnrootTree(BalancedTree(6))$edge)
@@ -264,42 +282,6 @@ test_that("CollapseNode() works", {
   expect_error(CollapseEdge(tree, 9))
 })
 
-test_that("DropTip() works", {
-  bal8 <- BalancedTree(8)
-  expect_equal(NTip(DropTip(bal8, 1:8)), 0)
-  expect_warning(expect_true(all.equal(bal8, DropTip(bal8, -1))))
-  expect_warning(expect_true(all.equal(bal8, DropTip(bal8, 99))))
-  expect_warning(expect_true(all.equal(bal8, DropTip(bal8, 'MissingTip'))))
-  expect_error(DropTip(bal8, list('Invalid format')))
-
-  expect_equal(DropTip(bal8, 7:8), DropTip(bal8, 15L))
-  expect_true(all.equal(ape::drop.tip(bal8, 6:8), DropTip(bal8, 6:8)))
-  expect_true(all.equal(ape::drop.tip(bal8, c(3, 5, 7)), DropTip(bal8, c(3, 5, 7))))
-
-  expect_equal(DropTip(Preorder(nasty), c(1, 3)),
-               Preorder(DropTip(nasty, c(1, 3))))
-
-
-  bigTree <- RandomTree(1284)
-  set.seed(1284)
-  bigTip <- sample(1:1284, 608)
-  expect_true(all.equal(drop.tip(bigTree, bigTip), DropTip(bigTree, bigTip)))
-  #microbenchmark(ape::drop.tip(bigTree, bigTip), DropTip(bigTree, bigTip), times = 25)
-  #profvis(replicate(25, DropTip(bigTree, bigTip)), interval = 0.005)
-})
-
-test_that("KeepTip() works", {
-  expect_warning(expect_true(all.equal(
-    BalancedTree(paste0('t', 5:8)),
-    KeepTip(BalancedTree(8), paste0('t', 5:9))
-  )))
-
-  expect_warning(expect_true(all.equal(
-    BalancedTree(paste0('t', 5:8)),
-    KeepTip(BalancedTree(8), 5:9)
-  )))
-})
-
 test_that("Binarification is uniform", {
   set.seed(0)
   Test <- function (tree, nTree, nSamples = 200L, ape = FALSE) {
@@ -334,7 +316,7 @@ test_that("LeafLabelInterchange() works", {
 
   abcd <- letters[1:4]
   sapply(1 + seq_len(100), function (i) {
-    # Check all pertubations
+    # Check all perturbations
     set.seed(i)
     expect_false(any(abcd == LeafLabelInterchange(BalancedTree(abcd), 4)$tip))